project: Move ES6/JSHint auto config to EmberProjectComponent class
